[FIX] UI: rename Instruction to State, update Progress factory descriptions.
